ollama-langchain-main/requirements.txt
streamlit
langchain
langchain-community
chromadb
pypdf
fastembed
tqdm
unstructured
pymupdf
python-docx
nomic
pytesseract 
pillow
ollama
pdfminer.six
